requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
djangorestframework==3.12.1
idna==2.10
lxml==4.5.2
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10

